separate worker class to single file worker.py
